# Conflicts:
#	src/components/school/design/SchoolDesign.jsx
#	src/components/school/design/dialog/FeedbackReportPopup.jsx

+++ src/services/AccountService.jsx
@@ -1,8 +1,12 @@
 import axiosClient from "../configs/APIConfig.jsx";
 
 export const signout = async () => {
-    localStorage.clear();
-    sessionStorage.clear();
+    if (localStorage.length > 0){
+        localStorage.clear();
+    }
+    if(sessionStorage.length > 0){
+        sessionStorage.clear()
+    }
     const response = await axiosClient.post("/account/logout");
     return response || null
 }
